+++ android/src/main/java/jp/shiguredo/react/webrtckit/WebRTCPeerConnectionObserver.java
@@ -212,12 +212,8 @@ final class WebRTCPeerConnectionObserver implements PeerConnection.Observer {
         sendDeviceEvent("peerConnectionAddedReceiver", params);
     }
 
-    // TODO: rtpReceiverがremoveされたタイミングで呼び出されるobserverが現状存在しないため未実装になっている
-    //       別口で調べてみたところ以下のコメントを見つけた
-    //       > (shino): Unified plan に onRemoveTrack が来たらこっち = void onTrack(RtpTransceiver transceiver) で対応する。
-    //       > 今は SDP semantics に関わらず onAddStream/onRemoveStream でシグナリングに通知している
-    //       現在libwebrtcのAndroid側実装にonRemoveTrackが存在しないため上記の様なワークアラウンドをするしかないようだが、
-    //       Streamに複数Trackが追加されている場合、onRemoveStreamを経由しないため、やはりすべてのケースには対応できない
+    // NOTE: libwebrtc の Android 側実装は onRemoveTrack に対応していないため、 https://github.com/shiguredo/shiguredo-webrtc-build でパッチを当てて対応しています。
+    //       上記のパッチが当たっていない AAR でもビルドを成功させるために、この API には意図的に @Override アノテーションを付与していません。
     public void onRemoveTrack(RtpReceiver receiver) {
         if (peerConnectionPair == null) return;
         Log.d("WebRTCModule", "onRemoveTrack()[" + peerConnectionPair.first + "] - receiver=" + receiver);