+++ CHANGES.md
@@ -11,6 +11,8 @@
 
 ## develop
 
+- [ADD] Android の onRemoveTrack に対応
+    - @enm10k
 - [CHANGE]  サポートする React Native の最低バージョンを 0.61.5 に更新
     - @enm10k