+++ CHANGES.md
@@ -11,6 +11,8 @@
 
 ## develop
 
+- [CHANGE] CircleCI の設定を削除する
+    - @kdxu
 - [CHANGE]  サポートする React Native の最低バージョンを 0.61.5 に更新
     - @enm10k